migrerer endre-kontaktinformasjon-rest-resource til nytt postresource-oppsett
