--- src/api/brukerprofil/endre-navkontaktinformasjon-api.ts
@@ -1,18 +1,0 @@
-import { apiBaseUri } from '../config';
-import { post } from '../api';
-
-export interface Request {
-    mobil?: Telefon;
-    jobb?: Telefon;
-    hjem?: Telefon;
-}
-
-interface Telefon {
-    retningsnummer: string;
-    identifikator: string;
-}
-
-export function postEndreNavKontaktinformasjon(request: Request, fødselsnummer: string): Promise<{}> {
-    const uri = `${apiBaseUri}/brukerprofil/${fødselsnummer}/telefonnummer/`;
-    return post(uri, request);
-}
