+++ src/redux/restReducers/brukerprofil/endreKontaktinformasjonRequest.ts
@@ -1,0 +1,10 @@
+export interface EndreKontaktinformasjonRequest {
+    mobil?: Telefon;
+    jobb?: Telefon;
+    hjem?: Telefon;
+}
+
+interface Telefon {
+    retningsnummer: string;
+    identifikator: string;
+}
